JENKINS-63071 move unit test to IT test to ease running it under java 11
